Refactor `issuerInfo` to `registrationCertificate` in Credential Issuer metadata
Currently, `issuerInfo` is being created in the TO of issuer metadata instead of it
being a field in the domain obj.
In the domain object now there is a value class `Wrprc` which exposes only a factory
method for the creation of the domain object.
Also, initialization of multiple `wrprc` has been removed.
