+++ README.md
@@ -549,7 +549,7 @@ Variable: `ISSUER_NONCE_ENCRYPTION_KEY_PASSWORD`
 Description: Password of the EC key-pair for nonce encryption.  
 Default value: N/A  
 
-Variable: `ISSUER_VERIFIERINFO_RRREGISTRATIONCERTIFICATE_XX`  
+Variable: `ISSUER_WRPRC`  
 Description: The JWT formatted registration certificate of Credential Issuer.  
 Default value: N/A
