+++ docker-compose/docker-compose.yaml
@@ -140,7 +140,7 @@ services:
       - ISSUER_METADATA_DISPLAY_0_LOCALE=en
       - ISSUER_METADATA_DISPLAY_0_LOGO_URI=https://eudiw.dev/ic-logo.svg
       - ISSUER_METADATA_DISPLAY_0_LOGO_ALTERNATIVE_TEXT=EU Digital Identity Wallet Logo
-      - ISSUER_VERIFIERINFO_REGISTRATIONCERTIFICATE_0=
+      - ISSUER_WRPRC=
     networks:
       - default
