+++ src/main/resources/application-develop.properties
@@ -70,4 +70,4 @@ spring.r2dbc.url=r2dbc:pool:postgresql://pid_issuer:pid_issuer@localhost:5432/pi
 # Trust Configuration
 issuer.trust.service-url=
 
-issuer.issuerinfo.registrationcertificate[0]=eyJ4NWMiOlsiTUlJQkhEQ0J4S0FEQWdFQ0FoUmp6T3Z1VUNXNFVieXMyNWgvRndHZlN4SFR1akFLQmdncWhrak9QUVFEQWpBUE1RMHdDd1lEVlFRRERBUjBaWE4wTUI0WERUSTJNRGN4TURBNE1EZ3dOMW9YRFRJM01EY3hNREE0TURnd04xb3dEekVOTUFzR0ExVUVBd3dFZEdWemREQlpNQk1HQnlxR1NNNDlBZ0VHQ0NxR1NNNDlBd0VIQTBJQUJKZlY0dFJ5U1dybTdsVkFTQjk0bUJLTTR3dlBKRkV1S012VWhwcUlyY3JkMkduam1PejJJS3Vrb0tkZVRjbTkzczM5U2w1NHl4UTI0Q3NLTVlmM2t3c3dDZ1lJS29aSXpqMEVBd0lEUndBd1JBSWdhdFJ6eUVidklLWk9wNVlyUDlJYW5ScVcyRTk1dC9vM3F1cmV0OXVXY2hrQ0lIcndEQ3dwQlJKNk9LQUNrZmNGRVZvWmNiTTU4NGZpRjVzb2h1TXNLcjg3Il0sInR5cCI6InJjLXdycCtqd3QiLCJhbGciOiJFUzI1NiJ9.eyJuYW1lIjoiS290bGluIElzc3VlciBTaWduZXIgRGV2Iiwic3ViX2xuIjoiTmlzY3kiLCJzdWIiOiJMRUlFVS0xMjM0NTY3ODkiLCJjb3VudHJ5IjoiRVUiLCJyZWdpc3RyeV91cmkiOiJodHRwczovL3JlZ2lzdHJ5LmV4YW1wbGUuZXUiLCJzcnZfZGVzY3JpcHRpb24iOlt7ImxhbmciOiJlbiIsInZhbHVlIjoiQW4gaW1wbGVtZW50YXRpb24gb2YgYSBjcmVkZW50aWFsIGlzc3Vpbmcgc2VydmljZSwgYWNjb3JkaW5nIHRvIE9wZW5JZDRWQ0kgLSB2MS4wIn1dLCJlbnRpdGxlbWVudHMiOlsiaHR0cHM6Ly91cmkuZXRzaS5vcmcvMTk0NzUvRW50aXRsZW1lbnQvUElEX1Byb3ZpZGVyIiwiaHR0cHM6Ly91cmkuZXRzaS5vcmcvMTk0NzUvRW50aXRsZW1lbnQvUUVBQV9Qcm92aWRlciIsImh0dHBzOi8vdXJpLmV0c2kub3JnLzE5NDc1L0VudGl0bGVtZW50L1BVQl9FQUFfUHJvdmlkZXIiXSwicHJpdmFjeV9wb2xpY3kiOiJodHRwczovL2Rldi5pc3N1ZXItYmFja2VuZC5ldWRpdy5kZXYvcHJpdmFjeSIsImluZm9fdXJpIjoiaHR0cHM6Ly9kZXYuaXNzdWVyLWJhY2tlbmQuZXVkaXcuZGV2Iiwic3VwcG9ydF91cmkiOiJodHRwczovL2Rldi5rb3RsaW5Jc3N1ZXJTaWduZXIuY29tL3N1cHBvcnQiLCJzdXBlcnZpc29yeV9hdXRob3JpdHkiOnsiZW1haWwiOiJzdXBlcnZpc29yeUBhdXRob3JpdHkuZXhhbXBsZS5ldSIsInBob25lIjoiKzQ5MzAxMjM0NTY3IiwidXJpIjoiaHR0cHM6Ly9zdXBlcnZpc29yeS5hdXRob3JpdHkuZXhhbXBsZS5ldSJ9LCJwb2xpY3lfaWQiOlsiMC40LjAuMTk0NzUuMy4xIl0sImNlcnRpZmljYXRlX3BvbGljeSI6Imh0dHBzOi8vZXhhbXBsZS5ldS9jZXJ0aWZpY2F0ZS1wb2xpY3kiLCJpYXQiOjE3ODM2ODEwMjgsInN0YXR1cyI6eyJzdGF0dXNfbGlzdCI6eyJpZHgiOjkxMCwidXJpIjoiaHR0cHM6Ly9pc3N1ZXIuZXVkaXcuZGV2L3Rva2VuX3N0YXR1c19saXN0L0VVL2V1LmV1cm9wYS5lYy5ldWRpLnBpZC4xLzIwZWY5ZTk4LTg2NzYtNDViYy04OWJhLTE3MDhmMGYyMTIzOSJ9fSwicHJvdmlkZXNfYXR0ZXN0YXRpb25zIjpbeyJmb3JtYXQiOiJkYytzZC1qd3QiLCJtZXRhIjp7InZjdF92YWx1ZXMiOlsidXJuOmV1ZGk6cGlkOjEiLCJ1cm46ZXUuZXVyb3BhLmVjLmV1ZGk6bGVhcm5pbmc6Y3JlZGVudGlhbDoxIl19fSx7ImZvcm1hdCI6Im1zb19tZG9jIiwibWV0YSI6eyJkb2N0eXBlX3ZhbHVlIjoiZXUuZXVyb3BhLmVjLmV1ZGkucGlkLjEifX0seyJmb3JtYXQiOiJtc29fbWRvYyIsIm1ldGEiOnsiZG9jdHlwZV92YWx1ZSI6Im9yZy5pc28uMTgwMTMuNS4xLm1ETCJ9fV19.h3wtRbSVXVGJVblxwHLQQivFN-iQiyyIGLmbFoNHpC8GRcKjuseUhbatzS-PGh7eDBPOCwh0VSgotEtjTiYKzA
+issuer.wrprc=eyJ4NWMiOlsiTUlJQkhEQ0J4S0FEQWdFQ0FoUmp6T3Z1VUNXNFVieXMyNWgvRndHZlN4SFR1akFLQmdncWhrak9QUVFEQWpBUE1RMHdDd1lEVlFRRERBUjBaWE4wTUI0WERUSTJNRGN4TURBNE1EZ3dOMW9YRFRJM01EY3hNREE0TURnd04xb3dEekVOTUFzR0ExVUVBd3dFZEdWemREQlpNQk1HQnlxR1NNNDlBZ0VHQ0NxR1NNNDlBd0VIQTBJQUJKZlY0dFJ5U1dybTdsVkFTQjk0bUJLTTR3dlBKRkV1S012VWhwcUlyY3JkMkduam1PejJJS3Vrb0tkZVRjbTkzczM5U2w1NHl4UTI0Q3NLTVlmM2t3c3dDZ1lJS29aSXpqMEVBd0lEUndBd1JBSWdhdFJ6eUVidklLWk9wNVlyUDlJYW5ScVcyRTk1dC9vM3F1cmV0OXVXY2hrQ0lIcndEQ3dwQlJKNk9LQUNrZmNGRVZvWmNiTTU4NGZpRjVzb2h1TXNLcjg3Il0sInR5cCI6InJjLXdycCtqd3QiLCJhbGciOiJFUzI1NiJ9.eyJuYW1lIjoiS290bGluIElzc3VlciBTaWduZXIgRGV2Iiwic3ViX2xuIjoiTmlzY3kiLCJzdWIiOiJMRUlFVS0xMjM0NTY3ODkiLCJjb3VudHJ5IjoiRVUiLCJyZWdpc3RyeV91cmkiOiJodHRwczovL3JlZ2lzdHJ5LmV4YW1wbGUuZXUiLCJzcnZfZGVzY3JpcHRpb24iOlt7ImxhbmciOiJlbiIsInZhbHVlIjoiQW4gaW1wbGVtZW50YXRpb24gb2YgYSBjcmVkZW50aWFsIGlzc3Vpbmcgc2VydmljZSwgYWNjb3JkaW5nIHRvIE9wZW5JZDRWQ0kgLSB2MS4wIn1dLCJlbnRpdGxlbWVudHMiOlsiaHR0cHM6Ly91cmkuZXRzaS5vcmcvMTk0NzUvRW50aXRsZW1lbnQvUElEX1Byb3ZpZGVyIiwiaHR0cHM6Ly91cmkuZXRzaS5vcmcvMTk0NzUvRW50aXRsZW1lbnQvUUVBQV9Qcm92aWRlciIsImh0dHBzOi8vdXJpLmV0c2kub3JnLzE5NDc1L0VudGl0bGVtZW50L1BVQl9FQUFfUHJvdmlkZXIiXSwicHJpdmFjeV9wb2xpY3kiOiJodHRwczovL2Rldi5pc3N1ZXItYmFja2VuZC5ldWRpdy5kZXYvcHJpdmFjeSIsImluZm9fdXJpIjoiaHR0cHM6Ly9kZXYuaXNzdWVyLWJhY2tlbmQuZXVkaXcuZGV2Iiwic3VwcG9ydF91cmkiOiJodHRwczovL2Rldi5rb3RsaW5Jc3N1ZXJTaWduZXIuY29tL3N1cHBvcnQiLCJzdXBlcnZpc29yeV9hdXRob3JpdHkiOnsiZW1haWwiOiJzdXBlcnZpc29yeUBhdXRob3JpdHkuZXhhbXBsZS5ldSIsInBob25lIjoiKzQ5MzAxMjM0NTY3IiwidXJpIjoiaHR0cHM6Ly9zdXBlcnZpc29yeS5hdXRob3JpdHkuZXhhbXBsZS5ldSJ9LCJwb2xpY3lfaWQiOlsiMC40LjAuMTk0NzUuMy4xIl0sImNlcnRpZmljYXRlX3BvbGljeSI6Imh0dHBzOi8vZXhhbXBsZS5ldS9jZXJ0aWZpY2F0ZS1wb2xpY3kiLCJpYXQiOjE3ODM2ODEwMjgsInN0YXR1cyI6eyJzdGF0dXNfbGlzdCI6eyJpZHgiOjkxMCwidXJpIjoiaHR0cHM6Ly9pc3N1ZXIuZXVkaXcuZGV2L3Rva2VuX3N0YXR1c19saXN0L0VVL2V1LmV1cm9wYS5lYy5ldWRpLnBpZC4xLzIwZWY5ZTk4LTg2NzYtNDViYy04OWJhLTE3MDhmMGYyMTIzOSJ9fSwicHJvdmlkZXNfYXR0ZXN0YXRpb25zIjpbeyJmb3JtYXQiOiJkYytzZC1qd3QiLCJtZXRhIjp7InZjdF92YWx1ZXMiOlsidXJuOmV1ZGk6cGlkOjEiLCJ1cm46ZXUuZXVyb3BhLmVjLmV1ZGk6bGVhcm5pbmc6Y3JlZGVudGlhbDoxIl19fSx7ImZvcm1hdCI6Im1zb19tZG9jIiwibWV0YSI6eyJkb2N0eXBlX3ZhbHVlIjoiZXUuZXVyb3BhLmVjLmV1ZGkucGlkLjEifX0seyJmb3JtYXQiOiJtc29fbWRvYyIsIm1ldGEiOnsiZG9jdHlwZV92YWx1ZSI6Im9yZy5pc28uMTgwMTMuNS4xLm1ETCJ9fV19.h3wtRbSVXVGJVblxwHLQQivFN-iQiyyIGLmbFoNHpC8GRcKjuseUhbatzS-PGh7eDBPOCwh0VSgotEtjTiYKzA
