+++ src/test/kotlin/eu/europa/ec/eudi/pidissuer/port/input/GetDeferredCredentialTest.kt
@@ -79,7 +79,7 @@ class GetDeferredCredentialTest {
         }
 
     private val jwtResourcePath = "/eu/europa/ec/eudi/pidissuer/adapter/out/jose/x5c/registration-certificate.jwt"
-    private val loadIssuerInfo = loadResource(jwtResourcePath).readText().trim()
+    private val registrationCertificate = loadResource(jwtResourcePath).readText().trim()
 
     private fun metadata(credentialRequestEncryption: CredentialRequestEncryption = CredentialRequestEncryption.NotSupported) =
         CredentialIssuerMetaData(
@@ -91,13 +91,7 @@ class GetDeferredCredentialTest {
             credentialResponseEncryption = CredentialResponseEncryption.NotSupported,
             attestationIssuers = nonEmptyListOf(attestationIssuer),
             preferredClientStatusPeriod = PreferredClientStatusPeriod(400.days),
-            issuerInfo =
-                nonEmptyListOf(
-                    IssuerInfo(
-                        format = "registration_cert",
-                        data = loadIssuerInfo,
-                    ),
-                ),
+            registrationCertificate = Wrprc.of(registrationCertificate),
         )
 
     @Test
